Added in viewPager introduction fragments
