+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,16 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <!--
         <activity android:name=".LoginActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        -->
+        <activity android:name=".IntroActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -25,10 +34,6 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".OverviewActivity" />
         <activity android:name=".DetailsActivity" />
-        <activity
-            android:name=".SetBudgetActivity"
-            android:label="@string/title_activity_set_budget"
-            android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
 
 </manifest>
