+++ app/src/main/java/edu/ucsb/cs/cs185/bryannaphan/fundguy/SetBudgetFragment.java
@@ -1,0 +1,73 @@
+package edu.ucsb.cs.cs185.bryannaphan.fundguy;
+
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class SetBudgetFragment extends Fragment {
+
+    private static final String ARG_LAYOUT_RES_ID = "layoutResId";
+    private int layoutResId;
+
+    public static SetBudgetFragment newInstance(int layoutResId) {
+        SetBudgetFragment setBudgetFragment = new SetBudgetFragment();
+
+        Bundle args = new Bundle();
+        args.putInt(ARG_LAYOUT_RES_ID, layoutResId);
+        setBudgetFragment.setArguments(args);
+
+        return setBudgetFragment;
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        if (getArguments() != null && getArguments().containsKey(ARG_LAYOUT_RES_ID)) {
+            layoutResId = getArguments().getInt(ARG_LAYOUT_RES_ID);
+        }
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        View view = inflater.inflate(layoutResId, container, false);
+
+        // TODO: Make this actually set the amount!
+        final Button setBudgetButton = (Button) view.findViewById(R.id.setBudgetButton);
+        /*
+        setBudgetButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent myIntent = new Intent(getActivity(), MainActivity.class);
+                getActivity().startActivity(myIntent);
+            }
+        });
+*/
+        /*FloatingActionButton fab = (FloatingActionButton) view.findViewById(R.id.fab);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Snackbar.make(view, "Budget has been successfully set", Snackbar.LENGTH_SHORT)
+                        .setAction("Action", null).show();
+            }
+        });*/
+
+        return view;
+    }
+
+}
