+++ app/src/main/res/layout/fragment_login.xml
@@ -1,0 +1,52 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/white"
+    tools:context="edu.ucsb.cs.cs185.bryannaphan.fundguy.LoginFragment">
+
+    <LinearLayout
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:paddingBottom="@dimen/activity_vertical_margin"
+        android:paddingLeft="@dimen/activity_horizontal_margin"
+        android:paddingRight="@dimen/activity_horizontal_margin"
+        android:paddingTop="@dimen/activity_vertical_margin"
+        android:gravity="center_horizontal"
+        android:weightSum="1"
+        android:background="@color/white" >
+
+        <TextView
+            android:text="Welcome to"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/welcomeText"
+            android:layout_gravity="center"/>
+
+        <TextView
+            android:text="fundGuy"
+            android:gravity="center"
+            android:textSize="80dp"
+            android:textColor="@color/black"
+            android:layout_width="match_parent"
+            android:layout_height="80dp"
+            android:id="@+id/titleText"
+            android:layout_marginTop="0dp"
+            android:layout_marginBottom="0dp"
+            android:layout_weight="0.10" />
+
+            <ImageView
+            android:id="@+id/mainIcon"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:background="@drawable/fundguy"
+            android:layout_alignParentTop="true"
+            android:layout_centerHorizontal="true"
+            android:scaleY="1"
+            android:scaleX="1"
+            android:padding="70dp"
+            android:layout_marginBottom="5dp" />
+
+    </LinearLayout>
+</FrameLayout>
